Create README.md
This is the general setting for the a/b test: "you received an analytical task from an international online store. Your predecessor failed to complete it: they launched an A/B test and then quit. They left only the technical specifications and the test results".
In this case there are unproportional a and b groups, late registrations and not enough time to check hypotheses on some of the participants.